move student/teachers specific components to corresponding folder
