+++ src/components/students/index.js
@@ -1,0 +1,2 @@
+export { default as Badge } from "./Badge.svelte";
+export { default as SideMenu } from "./SideMenu.svelte";
