+++ src/components/teachers/index.js
@@ -1,2 +1,4 @@
+export { default as FilterBlock } from "./FilterBlock.svelte";
 export { default as Header } from "./Header.svelte";
+export { default as Search } from "./Search.svelte";
 export { default as SideBar } from "./SideBar.svelte";
