+++ src/routes/students/Backpack.svelte
@@ -1,6 +1,6 @@
 <script>
   import I18n from "i18n-js";
-  import { Badge } from "../../components";
+  import { Badge } from "../../components/students";
   import { getBadges } from "../../api";
 
   let badgesPromise = getBadges();
